Fixes the port_delete call of inexistent port issue
The Hyper-V Neutron Agent processes the ports periodically.
If a port has been removed (instance was removed), then the
port will be unbound.
If an RPC call is received for the method "port_delete" for
a port that was not processed yet or the port was already
unbound, a 'TypeError' exception occurs.

The method _get_network_vswitch_map_by_port_id now
returns (None, None), if the port was not found and the
exception is avoided.

Closes-Bug: #1438711

Change-Id: I4f8b66e5af6b0b033e51549989f78f86d14840be