Hyper-V: Fixes unsupported 'TRUNK' endpoint mode
Checks if the 'TRUNK' endpoint mode is in the
SupportedEndpointModes before setting the
DesiredEndpointMode. If it is not supported,
log an error instead and skip this step.

This way, the Hyper-V Neutron Agent won't be stuck
in exceptions while processing this port. It can
continue to bind other ports (ports on a flat
network won't encounter this issue).

Closes-Bug: #1422449

Change-Id: I823f9b88a533204b704b9eb28381022cc56a8e69
(cherry picked from commit 7301acf0b6460a840fd7394a6dc1488f3dd8c6a2)