Fixes Hyper-V Agent state reporting under scale scenarios
While the Hyper-V Agent is processing each rule (ACL) for the
port, and if the number of such rules are large, the looping
interval thread doesn't get a chance to send the state report
to the controller.
This is fixed by yielding after applying every rule to other
threads who is waiting to run.

Closes-Bug: #1438084

Change-Id: Ic35aba8f29972e0f14b4e666973bf37b198270c1