Hyper-V: Makes sure empty array rules are not processed
When the security groups is updated, the HyperVSecurityGroupDriver
will create new rules and remove old rules. When a rule is removed
from the security group, the new rules list will be empty, which
will cause a list index out of range exception while adding the
new rule list.

Makes sure empty rule lists are not processed.

Closes-Bug: #1470581

Change-Id: Iac962371d8283524012f1cffd06363719440230d