Hyper-V: Removes extra cases for ingress ICMP rule
Previously, if the egress rule with no protocol specified
was added, the HyperVSecurityGroupsDriver would create an
inbound ICMP rule, in order for ping reply to be received.

This is because ICMP rules cannot be added as stateful on
Hyper-V and ping is expected to work if the rule is egress.
This would also cause the VM to be ping-able, which isn't
desirable.

Removes the special cases where ingress ICMP rules are
created. If ping is desirable, both ingress and egress
rules will have to be added.

Change-Id: If10d1408bf0f3ed487e394edd8e0b0ea52070a48
Closes-Bug: #1470443