Adds security_group_updated method in HyperVSecurityGroupDriver
With the changes introduced in securitygroups_rpc with the
below commit hyperv code breaks.
https://review.openstack.org/#/c/147713/

Because of this, any changes in the security groups are not applied
by the HyperVSecurityGroupDriver because of a NotImplemented error.

Adds the security_group_updated method as noop in
HyperVSecurityGroupDriver, solving the issue.

Closes-Bug: #1512278

Change-Id: Ia5d6fca3c3a2f827ade850bb43fc7e434e934a22
(cherry picked from commit f432ec44a78c4871a48a669c2aa81b33350ff41a)