Create a new object instead of querying the default one
Currently, a WMI Setting Data (Port VLAN, Port Security) is created
by querying the default object. Under large scale environments,
this can lead to huge execution times, as the query takes longer to
complete. This can lead to late port binding and missed DHCP
requests, leaving the VM with no IP.

A solution for this is to create the needed object (.new()) instead
of querying the default one.

Change-Id: Ib492afe5387373a40d964e76fa6b4b0ee3b5666d
Partial-Bug: #1505196