SG: Fixes ICMP protocol for IPv6
When creating an ICMP security group rule for IPv6 in neutron, the
HyperVSecurityGroupsDriver expects to receive a rule having the protocol
"ipv6-icmp", but receives a rule having the protocol "icmp".

Because of this, the Driver will sucessfully add the security group rule, but with
the wrong protocol (it will create a rule with protocol "1" - ICMP, instead
of "58" - ICMPv6).

Change-Id: I8e101952bf433148b9bd04726699c9f5a989f0c1
Closes-Bug: #1588340