Avoid port delete race conditions
We are attempting to unbind a port when receiving port delete events
or RPC calls. This is prone to race conditions, as the according
methods are not synchronized.

For this reason, we get ValueErrors when attempting to update the
cached information: http://paste.openstack.org/show/523766/

While this seems harmless (besides polluting the log), it's safer
if we just put proper locks in place.

Change-Id: I854d04618eb9eabc4cee170bd21e98317ec15678
Closes-Bug: #1596916