Fixes port security rules deletion
Port security rules might contain rules that apply
to any protocol.

Currently there is no checking to see if a rule
which will be deleted belongs to a security
rule that applies to all protocols.

This patch addresses this issue by adding a check
to make sure the rule can be safely removed.

Closes-Bug: #1634082
Change-Id: Icfa7d77609394a2a7030f68c2643baf309a5de74