Adds trunked VLANs support
Instances can be spawned by passing in a port_id to nova.
Those ports can be trunk ports, and can have multiple
subports with different segmentation IDs. In this case,
the port will have to be set in Trunk Mode, setting the
trunk port's segmentation_id as the port's NativeVlanId,
and the trunk's subports' segmentation_ids as the
TrunkVlanIdArray.

Depends-On: I0de69f88cc52d8904579e18678e0be8d77bd9a76

Implements: blueprint hyperv-vlan-trunks

Change-Id: I1751e88ff16b142995f67cec7fa698c1a98587f2