Updates networkutils caches on worker error
os_win's networkutils caches switch ports, ACLs, VLAN SDs, etc.
for performance reasons, and it is refreshed on exception in the
daemon_loop, but that case is not happening anymore since we switched
to worker threads; the exceptions are treated by the workers.

Now, if a worker encounters an exception, it will set a refresh cache
flag, and the daemon_loop will refresh the caches if the flag is set.

Change-Id: Ia72ddaf0743bb6b0599f5021339da6cfe07d9924
Closes-Bug: #1676941
(cherry picked from commit 832ab6521bea53b60b4cb3a9196cafcfab816537)