Adds SR-IOV support
In order to use SR-IOV on a compute node, the feature has
to be enabled and supported by the OS and the NICs.

The nova-compute service will report the SR-IOV VFs as
PCI devices. SR-IOV is requested by creating a neutron port with
the binding:vnic_type direct.

The Hyper-V Mechanism Driver already supports ports with vnic_type direct,
so no changes are necessary to it.

An SR-IOV neutron port will contain the pci_slot in the port's profile.
When can use that information in order to determine whether SR-IOV has to
be enabled or not.

Depends-On: Ied27bb4e1b36e743b33c6b65351c2ee942ef535e

Implements: blueprint networking-hyperv-sriov

Change-Id: I131d65e45c59eec577bb8d076f674146e37de8e3