Removes SG rules on port with SG disabled
The SecurityGroupsDriver relies on its cache when removing a port's
ACL rules if the port's port_security_enabled field is False.

If the port was updated while the agent was down, the cache will be
empty, and thus skip removing the port's ACLs.

This patch addresses this issue by removing all of the port's ACLs if
the port's port_security_enabled is False.

Change-Id: Ibda80fbd17310e13ceb7fe4e6db7f68e6403e87b
Closes-Bug: #1747666
(cherry picked from commit 13a1d75fe79dc05c7822625b29fe07554855d70e)