Ignore missing ports when handling security groups
This issue affects the neutron OVS agent, when using the Hyper-V
security groups driver.

In quite a few situations, OVS ports may be leaked. This becomes
more troublesome as the OVS agent will try to update rules for
Hyper-V switch ports that do not exist anymore and fail.

Note that the agent handles multiple ports at a time. If one of
them is a leaked port, the others won't get processed.

The security groups driver may just ignore missing ports.

Closes-Bug: #1794975

Change-Id: Ic59ba7ca68828dcead99973fadc4ec798e116833